// src/store/Editor/mutations/mutation-type.ts
export enum MUTATION_TYPE {
  LOAD = 'LOAD',
  LOAD_BY_CACHE = 'LOAD_BY_CACHE',
  ADD_PAGE = 'ADD_PAGE',
  REMOVE_PAGE = 'REMOVE_PAGE',
  SELECT_PAGE = 'SELECT_PAGE',
  CHANGE_PAGE = 'CHANGE_PAGE',
  EDIT_PAGE = 'EDIT_PAGE',
  DELETE_PAGE = 'DELETE_PAGE',
  COPY_PAGE = 'COPY_PAGE',
  ADD_COMPONENT = 'ADD_COMPONENT',
  REMOVE_COMPONENT = 'REMOVE_COMPONENT',
  SELECT_COMPONENT = 'SELECT_COMPONENT',
  EXTRACT_COMPONENT = 'EXTRACT_COMPONENT',
  DELETE_EXTRACT_COMPONENT = 'DELETE_EXTRACT_COMPONENT',
  UNDO = 'UNDO',
  REDO = 'REDO',
  INIT = 'INIT',
  UPDATE_COMPONENT = 'UPDATE_COMPONENT',
  RESIZE = 'RESIZE',
  ADD_EVENT = 'ADD_EVENT',
  REMOVE_EVENT = 'REMOVE_EVENT',
  UPDATE_EVENT = 'UPDATE_EVENT',
  DRAG_COMPONENT = 'DRAG_COMPONENT',
  DRAG_TREE = 'DRAG_TREE',
  ENTER_CONTAINER = 'ENTER_CONTAINER',
  LEAVE_CONTAINER = 'LEAVE_CONTAINER',
}